winy/tools/requirements.txt
streamlit==1.54.0
ollama==0.6.1
qdrant-client==1.17.0
fastembed==0.7.4
psycopg2-binary==2.9.11
